chore: remove @modern-js/adapter-helpers
