--- packages/server/adapter-helpers/.eslintrc.js
@@ -1,8 +1,0 @@
-module.exports = {
-  root: true,
-  extends: ['@modern-js'],
-  parserOptions: {
-    tsconfigRootDir: __dirname,
-    project: ['./tsconfig.json'],
-  },
-};
