--- packages/server/adapter-helpers/CHANGELOG.md
@@ -1,242 +1,0 @@
-# @modern-js/adapter-helpers
-
-## 1.3.0
-
-### Minor Changes
-
-- 77a8e9e: feat: support bff operators
-
-## 1.2.6
-
-### Patch Changes
-
-- a1198d509: feat: bump babel 7.18.0
-
-## 1.2.5
-
-### Patch Changes
-
-- 51c4e2566: fix: isRelative function in babel-plugin-resolver plugin can't handle windows relative path correctly
-
-## 1.2.4
-
-### Patch Changes
-
-- d32f35134: chore: add modern/jest/eslint/ts config files to .npmignore
-
-## 1.2.3
-
-### Patch Changes
-
-- 6cffe99d: chore:
-  remove react eslint rules for `modern-js` rule set.
-  add .eslintrc for each package to speed up linting
-- 60f7d8bf: feat: add tests dir to npmignore
-- befd9e5b: fix: compatible with babel-plugin-resolver's handling of relative paths on windows
-- 3bf4f8b0: feat: support start api server only
-
-## 1.2.2
-
-### Patch Changes
-
-- 55e18278: chore: remove unused dependencies and devDependencies
-
-## 1.2.1
-
-### Patch Changes
-
-- 83166714: change .npmignore
-
-## 1.2.0
-
-### Minor Changes
-
-- cfe11628: Make Modern.js self bootstraping
-
-## 1.1.1
-
-### Patch Changes
-
-- 0fa83663: support more .env files
-
-## 1.1.0
-
-### Minor Changes
-
-- 96119db2: Relese v1.1.0
-
-## 1.0.0
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 30ac27c: feat: add generator package description
-- 0fd196e: feat: fix bugs
-- 204c626: feat: initial
-- 63be0a5: fix: #118 #104
-
-## 1.0.0-rc.23
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 30ac27c: feat: add generator package description
-- 0fd196e: feat: fix bugs
-- 204c626: feat: initial
-- 63be0a5: fix: #118 #104
-
-## 1.0.0-rc.22
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 30ac27c: feat: add generator package description
-- 0fd196e: feat: fix bugs
-- 204c626: feat: initial
-- 63be0a5: fix: #118 #104
-
-## 1.0.0-rc.21
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 30ac27c: feat: add generator package description
-- 0fd196e: feat: fix bugs
-- 204c626: feat: initial
-- 63be0a5: fix: #118 #104
-
-## 1.0.0-rc.20
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 30ac27c: feat: add generator package description
-- feat: fix bugs
-- 204c626: feat: initial
-- 63be0a5: fix: #118 #104
-
-## 1.0.0-rc.19
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 30ac27c: feat: add generator package description
-- 204c626: feat: initial
-- 63be0a5: fix: #118 #104
-
-## 1.0.0-rc.18
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 30ac27c: feat: add generator package description
-- 204c626: feat: initial
-- 63be0a5: fix: #118 #104
-
-## 1.0.0-rc.17
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 30ac27c: feat: add generator package description
-- 204c626: feat: initial
-- fix: #118 #104
-
-## 1.0.0-rc.16
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 30ac27c: feat: add generator package description
-- 204c626: feat: initial
-
-## 1.0.0-rc.15
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 30ac27c: feat: add generator package description
-- 204c626: feat: initial
-
-## 1.0.0-rc.14
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 204c626: feat: initial
-
-## 1.0.0-rc.13
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 204c626: feat: initial
-
-## 1.0.0-rc.12
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 204c626: feat: initial
-
-## 1.0.0-rc.11
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 204c626: feat: initial
-
-## 1.0.0-rc.10
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 204c626: feat: initial
-
-## 1.0.0-rc.9
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 204c626: feat: initial
-
-## 1.0.0-rc.8
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 204c626: feat: initial
-
-## 1.0.0-rc.7
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 204c626: feat: initial
-
-## 1.0.0-rc.6
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 204c626: feat: initial
-
-## 1.0.0-rc.5
-
-### Patch Changes
-
-- 224f7fe: fix server route match
-- 204c626: feat: initial
-
-## 1.0.0-rc.4
-
-### Patch Changes
-
-- fix server route match
-- 204c626: feat: initial
-
-## 1.0.0-rc.3
-
-### Patch Changes
-
-- feat: initial
