--- packages/server/adapter-helpers/jest.config.js
@@ -1,7 +1,0 @@
-const sharedConfig = require('@scripts/jest-config');
-
-/** @type {import('@jest/types').Config.InitialOptions} */
-module.exports = {
-  ...sharedConfig,
-  rootDir: __dirname,
-};
