--- packages/server/adapter-helpers/modern.config.js
@@ -1,2 +1,0 @@
-/** @type {import('@modern-js/module-tools').UserConfig} */
-module.exports = {};
