--- packages/server/adapter-helpers/package.json
@@ -1,77 +1,0 @@
-{
-  "name": "@modern-js/adapter-helpers",
-  "description": "The meta-framework suite designed from scratch for frontend-focused modern web development.",
-  "homepage": "https://modernjs.dev",
-  "bugs": "https://github.com/modern-js-dev/modern.js/issues",
-  "repository": "modern-js-dev/modern.js",
-  "license": "MIT",
-  "keywords": [
-    "react",
-    "framework",
-    "modern",
-    "modern.js"
-  ],
-  "version": "1.3.0",
-  "jsnext:source": "./src/index.ts",
-  "types": "./src/index.ts",
-  "main": "./dist/js/node/index.js",
-  "module": "./dist/js/treeshaking/index.js",
-  "jsnext:modern": "./dist/js/modern/index.js",
-  "exports": {
-    ".": {
-      "jsnext:source": "./src/index.ts",
-      "default": "./dist/js/node/index.js"
-    }
-  },
-  "scripts": {
-    "prepare": "pnpm build",
-    "prepublishOnly": "only-allow-pnpm && pnpm build -- --platform",
-    "new": "modern new",
-    "build": "wireit",
-    "test": "wireit"
-  },
-  "dependencies": {
-    "@babel/runtime": "^7.18.0"
-  },
-  "devDependencies": {
-    "@scripts/build": "workspace:*",
-    "@scripts/jest-config": "workspace:*",
-    "@types/jest": "^27",
-    "@types/node": "^14",
-    "jest": "^27",
-    "typescript": "^4"
-  },
-  "modernConfig": {
-    "output": {
-      "packageMode": "node-js"
-    }
-  },
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org/",
-    "access": "public",
-    "types": "./dist/types/index.d.ts"
-  },
-  "wireit": {
-    "build": {
-      "command": "modern build",
-      "files": [
-        "src/**/*",
-        "tsconfig.json",
-        "package.json"
-      ],
-      "output": [
-        "dist/**/*"
-      ]
-    },
-    "test": {
-      "command": "jest --passWithNoTests",
-      "files": [
-        "src/**/*",
-        "tsconfig.json",
-        "package.json",
-        "tests/**/*"
-      ],
-      "output": []
-    }
-  }
-}
