--- packages/server/adapter-helpers/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "@modern-js/tsconfig/base",
-  "compilerOptions": {
-    "declaration": false,
-    "jsx": "preserve",
-    "baseUrl": "./",
-    "isolatedModules": true
-  },
-  "include": ["src"]
-}
