+++ packages/server/bff-core/src/index.ts
@@ -2,6 +2,11 @@ export { Api } from './api';
 export { HttpError, ValidationError } from './errors/http';
 export * from './router';
 export * from './types';
-export * from './utils';
 export * from './client';
 export * from './operators/http';
+export {
+  getRelativeRuntimePath,
+  HANDLER_WITH_META,
+  isWithMetaHandler,
+  createStorage,
+} from './utils';
