+++ packages/server/bff-core/src/utils/debug.ts
@@ -1,0 +1,3 @@
+import { createDebugger } from '@modern-js/utils';
+
+export const debug = createDebugger('bff');
