+++ packages/server/bff-core/src/utils/meta.ts
@@ -1,0 +1,5 @@
+export const HANDLER_WITH_META = 'HANDLER_WITH_META';
+
+export const isWithMetaHandler = (handler: any) => {
+  return typeof handler === 'function' && handler[HANDLER_WITH_META];
+};
