+++ packages/server/plugin-egg/src/context.ts
@@ -1,4 +1,4 @@
-import { createStorage } from '@modern-js/adapter-helpers';
+import { createStorage } from '@modern-js/bff-core';
 import type { Context } from 'egg';
 
 const { run, useContext } = createStorage<Context>();
