+++ packages/server/plugin-express/package.json
@@ -41,7 +41,6 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.18.0",
-    "@modern-js/adapter-helpers": "workspace:^1.3.0",
     "@modern-js/bff-core": "workspace:^1.2.2",
     "@modern-js/bff-runtime": "workspace:^1.4.0",
     "@modern-js/types": "workspace:^1.6.2",
