+++ packages/server/plugin-nest/package.json
@@ -43,7 +43,6 @@
     "test": "wireit"
   },
   "dependencies": {
-    "@modern-js/adapter-helpers": "workspace:^1.3.0",
     "@modern-js/bff-core": "workspace:^1.2.2",
     "@modern-js/bff-runtime": "workspace:^1.4.0",
     "@modern-js/utils": "workspace:^1.9.0",
